Ejemplos/RaspberryPi/Comunicacion/MQTT_Publish/requirements.txt
# Comunicación MQTT
paho-mqtt==1.6.1

# Sensores
RPi.GPIO==0.7.1
Adafruit-DHT==1.4.0

# Utilidades
python-dotenv==1.0.0
schedule==1.2.0 